% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaguetable.R
\name{leaguetable}
\alias{leaguetable}
\title{Create a league table}
\usage{
leaguetable(dataset, home, away, score_home, score_away, date, date_start,
  date_end, matchday, matchday_start, matchday_end, points = c(3, 1, 0),
  rank_by = c("GD", "GF"), HA_tables = FALSE)
}
\arguments{
\item{dataset}{A dataset with the results.}

\item{home}{The name of the home team column in the dataset as a character string.}

\item{away}{The name of the away team column in the dataset  as a character string.}

\item{score_home}{The name of the home team goals column in the dataset as a character string.}

\item{score_away}{The name of the away team goals column in the dataset as a character string.}

\item{date}{The name of the date column in the dataset as a character string (optional).}

\item{date_start}{The earliest date to include if not the earliest date in the dataset as a character string in the format "YY-mm-dd" (optional).}

\item{date_end}{The last date to include if not the last date in the dataset as character string in the format "YY-mm-dd" (optional).}

\item{matchday}{The name of the matchday column in the dataset as a character string (optional).}

\item{matchday_start}{The earliest matchday to include if not the earliest in the dataset as an integer (optional).}

\item{matchday_end}{The last matchday to include if not the last in the dataset as an integer (optional).}

\item{points}{A vector of integers of length three containing the points awarded for wins, draws and losses. Defaults to c(3,1,0).}

\item{rank_by}{A character vector with the order of arguments to sort the league table following "Pts". Possible arguments are "GD", "GF", "GA", "W","D","L", Defaults to c("GD","GF").}

\item{HA_table}{Logical value to indicate whether home and away results should be displayed in the table. Defaults to FALSE.}

\item{DC}{Logical value to indicate whether direct comparison of teams with equal points should be applied. The ranking of teams with equal numbers of points use the same order as defined in rank_by.}
}
\value{
A data frame in the form of a league table.
}
\description{
Create a league table from individually defined criteria from a results dataset.
}
\details{
Mandatory parameters are a dataset with match results as well the names of the variables for home and away team and their respective scored goals.

Optional parameters are the point distribution that will be awarded (win, draw, lose), the order of ranking criteria to use (Pts = Points, GD = Goal difference, GF = Goals for),
the display of additional columns with separate home and away tables and the dates and matchdays to be used for calculation.

If a date range and a matchday range are set, the subset of matches that fit both selection criteria will be used for calculation.
}
\examples{
#create league table for La Liga 1994/95 with two points rewards with home and away results
library(engsoccerdata)
leaguetable(dataset=engsoccerdata::spain[which(engsoccerdata::spain$Season==1994),], home="home", away="visitor", score_home="hgoal", score_away="vgoal", date="Date", points = c(2,1,0), HA_tables = TRUE)
}
