% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TouRnament.R
\docType{package}
\name{TouRnament}
\alias{TouRnament}
\alias{TouRnament-package}
\title{TouRnament Package}
\description{
Package with Functions Related to Sports Competitions
}
\author{
Tobias Wolfanger \email{tobias.wolfanger@gmx.de}
}
